Moved DateFragment to TodoTxt
